Merge de la rama feature/project-structure hacia la rama master

+++ LiquidacionPeajesNew.Application/DTOs/Responses/GaritaResponse.cs
@@ -5,7 +5,7 @@
         public long IdGarita { get; set; }
         public string NombreGarita { get; set; }
         public string RucProveedor { get; set; }
-        public short IdZonaGarita { get; set; }
+        public byte IdZonaGarita { get; set; }
         public string NombreZonaGarita { get; set; }
         public int UbicacionGarita { get; set; }
         public bool IdaVuelta { get; set; }
