+++ LiquidacionPeajesNew.Application/Services/ZonaGaritaService/IZonaGaritaService.cs
@@ -6,9 +6,9 @@ namespace LiquidacionPeajesNew.Application.Services.ZonaGaritaService
     public interface IZonaGaritaService
     {
         Task<ApiResponse<IEnumerable<ZonaGaritaResponse>>> GetAllAsync();
-        Task<ApiResponse<ZonaGaritaResponse>> GetByIdAsync(short id);
-        Task<ApiResponse<short>> AddAsync(ZonaGaritaRequest request);
-        Task<ApiResponse<short>> UpdateAsync(ZonaGaritaRequest request);
-        Task<ApiResponse<short>> DeleteAsync(short id);
+        Task<ApiResponse<ZonaGaritaResponse>> GetByIdAsync(byte id);
+        Task<ApiResponse<byte>> AddAsync(ZonaGaritaRequest request);
+        Task<ApiResponse<byte>> UpdateAsync(ZonaGaritaRequest request);
+        Task<ApiResponse<byte>> DeleteAsync(byte id);
     }
 }
