+++ LiquidacionPeajesNew.Application/Services/ZonaGaritaService/ZonaGaritaService.cs
@@ -30,7 +30,7 @@ namespace LiquidacionPeajesNew.Application.Services.ZonaGaritaService
             );
         }
 
-        public async Task<ApiResponse<ZonaGaritaResponse>> GetByIdAsync(short id)
+        public async Task<ApiResponse<ZonaGaritaResponse>> GetByIdAsync(byte id)
         {
             var entity = await _repository.GetByIdAsync(id);
             if (entity.IdZonaGarita == 0)
@@ -51,14 +51,14 @@ namespace LiquidacionPeajesNew.Application.Services.ZonaGaritaService
             );
         }
 
-        public async Task<ApiResponse<short>> AddAsync(ZonaGaritaRequest request)
+        public async Task<ApiResponse<byte>> AddAsync(ZonaGaritaRequest request)
         {
             var mapped = _mapper.Map<ZonaGaritaEntity>(request);
 
             var existe = await _repository.ExistsAsync(mapped.IdZonaGarita, mapped.ZonaGarita);
             if (existe)
             {
-                return new ApiResponse<short>(
+                return new ApiResponse<byte>(
                     status: false,
                     value: 0,
                     messageCode: AppResponseCode.RecordAlreadyExists
@@ -67,21 +67,21 @@ namespace LiquidacionPeajesNew.Application.Services.ZonaGaritaService
 
             await _repository.AddAsync(mapped);
 
-            return new ApiResponse<short>(
+            return new ApiResponse<byte>(
                 status: true,
                 value: mapped.IdZonaGarita,
                 messageCode: AppResponseCode.OperationCompletedSuccessfully
             );
         }
 
-        public async Task<ApiResponse<short>> UpdateAsync(ZonaGaritaRequest request)
+        public async Task<ApiResponse<byte>> UpdateAsync(ZonaGaritaRequest request)
         {
             var mapped = _mapper.Map<ZonaGaritaEntity>(request);
 
             var existe = await _repository.ExistsAsync(mapped.IdZonaGarita, mapped.ZonaGarita);
             if (existe)
             {
-                return new ApiResponse<short>(
+                return new ApiResponse<byte>(
                     status: false,
                     value: 0,
                     messageCode: AppResponseCode.RecordAlreadyExists
@@ -90,19 +90,19 @@ namespace LiquidacionPeajesNew.Application.Services.ZonaGaritaService
 
             await _repository.UpdateAsync(mapped);
 
-            return new ApiResponse<short>(
+            return new ApiResponse<byte>(
                 status: true,
                 value: mapped.IdZonaGarita,
                 messageCode: AppResponseCode.OperationCompletedSuccessfully
             );
         }
 
-        public async Task<ApiResponse<short>> DeleteAsync(short id)
+        public async Task<ApiResponse<byte>> DeleteAsync(byte id)
         {
             var entity = await _repository.GetByIdAsync(id);
             if (entity.IdZonaGarita == 0)
             {
-                return new ApiResponse<short>(
+                return new ApiResponse<byte>(
                     status: false,
                     value: id,
                     messageCode: AppResponseCode.RecordNotFoundInDatabase
@@ -111,7 +111,7 @@ namespace LiquidacionPeajesNew.Application.Services.ZonaGaritaService
 
             await _repository.DeleteAsync(id);
 
-            return new ApiResponse<short>(
+            return new ApiResponse<byte>(
                 status: true,
                 value: entity.IdZonaGarita,
                 messageCode: AppResponseCode.OperationCompletedSuccessfully
